Change value in leaderboards from int to double

+++ noteblock-api/src/main/java/gg/makera/noteblock/api/NoteblockAPI.java
@@ -20,7 +20,7 @@ public interface NoteblockAPI extends AutoCloseable {
     CompletableFuture<LeaderboardUpdateResponse> updateLeaderboard(int serverId,
                                                                    @NotNull String leaderboardId,
                                                                    @NotNull String playerName,
-                                                                   int value);
+                                                                   double value);
 
     CompletableFuture<UserInfoResponse> getUserInfo();
